fix display, clean up tests and reorganize
