+++ tests/integration/README.md
@@ -1,0 +1,12 @@
+# Integration Tests
+
+This directory is for integration tests that verify how different components of the application work together.
+
+Integration tests should test complete features or workflows rather than individual units of code.
+
+Currently, there are no integration tests, but they should be added here as needed. Some good candidates for 
+integration tests would include:
+
+1. End-to-end tests of CLI commands
+2. Tests requiring actual network communication (behind feature flags)
+3. Tests that verify multiple components work together correctly
